requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-jwt-extended==4.5.3
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
twilio==8.10.0
numpy>=1.26.4
opencv-python-headless>=4.9.0.80
Pillow>=10.3.0
joblib>=1.3.2
scikit-learn>=1.4.0
gunicorn==21.2.0
resend
psycopg2-binary
xgboost==2.0.3
cloudinary
anthropic
ultralytics
timm
torch
torchvision